Add and use verify_copy function in jutil lib
The HecDSSUtilities.copyFile function is expected to return an integer
indicating success/failure of the operation, but this was determined to
be unreliable.  Tested by removing write permissions to a DSS file and
monitoring integer upon failure -- still returned success value (0).

The verify_copy function provides a manual method to verify success of
the operation by testing if every pathname in the source DSS file exists
in the destination DSS file.